// frontend/src/lib/api/models/adminRecordScheduleSearchRequestPlatformType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Stream Archive API
 * 멀티 플랫폼 스트리밍 녹화 시스템 API
 * OpenAPI spec version: 1.0.0
 */

export type AdminRecordScheduleSearchRequestPlatformType =
  (typeof AdminRecordScheduleSearchRequestPlatformType)[keyof typeof AdminRecordScheduleSearchRequestPlatformType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminRecordScheduleSearchRequestPlatformType = {
  CHZZK: "CHZZK",
  TWITCH: "TWITCH",
  SOOP: "SOOP",
} as const;