// frontend/src/lib/api/models/adminRecordScheduleSearchRequestScheduleType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Stream Archive API
 * 멀티 플랫폼 스트리밍 녹화 시스템 API
 * OpenAPI spec version: 1.0.0
 */

export type AdminRecordScheduleSearchRequestScheduleType =
  (typeof AdminRecordScheduleSearchRequestScheduleType)[keyof typeof AdminRecordScheduleSearchRequestScheduleType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminRecordScheduleSearchRequestScheduleType = {
  ONCE: "ONCE",
  ALWAYS: "ALWAYS",
  N_DAYS_OF_EVERY_WEEK: "N_DAYS_OF_EVERY_WEEK",
  SPECIFIC_DAY: "SPECIFIC_DAY",
} as const;